requirements.txt
flask
python-dotenv
google-generativeai
gspread
oauth2client